Jump to number animation added
